+++ src/main/java/server/ServerChatFusion.java
@@ -1,9 +1,8 @@
 package main.java.server;
 
 import main.java.OpCode;
-import main.java.reader.Message;
-import main.java.reader.MessageReader;
-import main.java.reader.Reader;
+import main.java.Utils.RequestFactory;
+import main.java.reader.*;
 
 import java.io.IOException;
 import java.io.UncheckedIOException;
@@ -22,16 +21,20 @@ public class ServerChatFusion {
     private static final int BUFFER_SIZE = 1024;
     private static final Logger logger = Logger.getLogger(ServerChatFusion.class.getName());
     private static final long TIMEOUT = 60_000;
+    private final String serverName;
     private final ServerSocketChannel serverSocketChannel;
     private final Selector selector;
     private final Thread console;
     private final StateController stateController = new StateController();
-
     private final ServerChatFusion leader = this;
-
     private final HashMap<String, SelectionKey> clientConnected = new HashMap<>();
 
-    public ServerChatFusion(int port) throws IOException {
+    public ServerChatFusion(String serverName, int port) throws IOException {
+        Objects.requireNonNull(serverName);
+        if (serverName.isEmpty()) {
+            throw new IllegalArgumentException();
+        }
+        this.serverName = serverName;
         serverSocketChannel = ServerSocketChannel.open();
         serverSocketChannel.bind(new InetSocketAddress(port));
         selector = Selector.open();
@@ -40,15 +43,16 @@ public class ServerChatFusion {
     }
 
     public static void main(String[] args) throws NumberFormatException, IOException {
-        if (args.length != 1) {
+        if (args.length != 2) {
             usage();
             return;
         }
-        new ServerChatFusion(Integer.parseInt(args[0])).launch();
+
+        new ServerChatFusion(args[0], Integer.parseInt(args[1])).launch();
     }
 
     private static void usage() {
-        System.out.println("Usage : ServerSumBetter port");
+        System.out.println("Usage : ServerSumBetter serverName port");
     }
 
     private boolean isLeader() {
@@ -156,12 +160,29 @@ public class ServerChatFusion {
     }
 
     /**
-     * Add a message to all connected clients queue
+     * Add a request to all connected clients queue
      *
-     * @param msg The message to broadcast to every clients
+     * @param request The request to broadcast to every client
      */
-    private void broadcast(Message msg) {
-        clientConnected.entrySet().stream().filter(client -> !client.getKey().equals(msg.login())).map(client -> (Context) client.getValue().attachment()).forEach(context -> context.queueMessage(msg));
+    private void broadcast(Request request) {
+        clientConnected.values().stream().map(key -> (Context) key.attachment()).forEach(context -> context.queueRequest(request));
+
+        //TODO send to servers
+    }
+
+    public void addClient(String login, SelectionKey key) {
+        var duplicate = (clientConnected.putIfAbsent(login, key) != null);
+
+        var client = (Context) key.attachment();
+
+        if (duplicate) {
+            // send connection refused
+            client.queueRequest(RequestFactory.loginRefused());
+            return;
+        }
+
+        // send connection accept
+        client.queueRequest(RequestFactory.loginAccepted(serverName));
     }
 
     private enum State {
@@ -190,13 +211,13 @@ public class ServerChatFusion {
         private final SocketChannel sc;
         private final ByteBuffer bufferIn = ByteBuffer.allocate(BUFFER_SIZE);
         private final ByteBuffer bufferOut = ByteBuffer.allocate(BUFFER_SIZE);
-        private final MessageReader messageReader = new MessageReader();
-        private final ArrayDeque<Message> queue = new ArrayDeque<>();
-        private final ServerChatFusion server; // we could also have Context as an instance class
 
+        private final MessageReader messageReader = new MessageReader();
+        private final StringReader stringReader = new StringReader();
+        private final ArrayDeque<Request> requestQueue = new ArrayDeque<>();
+        private final ServerChatFusion server; // we could also have Context as an instance class
         private OpCode watcher = OpCode.IDLE;
         private boolean activeSinceLastTimeoutCheck = true;
-        // which would naturally give access to ServerChatInt.this
         private boolean closed = false;
 
         private Context(ServerChatFusion server, SelectionKey key) {
@@ -212,16 +233,43 @@ public class ServerChatFusion {
          * after the call
          */
         private void processIn() {
+            System.out.println("Bonjour !");
             if (watcher == OpCode.IDLE) {
                 var optionalWatcher = OpCode.getOpCodeFromInt(bufferIn.flip().getInt());
+                bufferIn.compact();
                 if (optionalWatcher.isPresent()) {
                     watcher = optionalWatcher.get();
                 } else {
-                    // Close the connection if it send a wrong OpCode
+                    // Close the connection if it sent a wrong OpCode
                     silentlyClose();
                 }
             }
 
+
+            switch (watcher) {
+                case LOGIN_ANONYMOUS -> {
+                    var status = stringReader.process(bufferIn, 30);
+                    switch (status) {
+                        case DONE -> {
+                            var login = stringReader.get();
+                            server.addClient(login, key);
+                            stringReader.reset();
+                        }
+                        case ERROR -> {
+                            //TODO connection refused, not silently close
+                            silentlyClose();
+                        }
+                        case REFILL -> {
+                        }
+                    }
+                }
+                default -> {
+                    // TODO temporary
+                    throw new UnsupportedOperationException();
+                }
+            }
+
+            /*
             for (; ; ) {
                 Reader.ProcessStatus status = messageReader.process(bufferIn);
                 switch (status) {
@@ -239,15 +287,16 @@ public class ServerChatFusion {
                     }
                 }
             }
+             */
         }
 
         /**
-         * Add a message to the message queue, tries to fill bufferOut and updateInterestOps
+         * Add a request to the request queue, tries to fill bufferOut and updateInterestOps
          *
-         * @param msg
+         * @param request request containing the opcode of the request and a buffer with the request's content
          */
-        public void queueMessage(Message msg) {
-            queue.add(msg);
+        public void queueRequest(Request request) {
+            requestQueue.add(request);
             processOut();
             updateInterestOps();
         }
@@ -256,11 +305,11 @@ public class ServerChatFusion {
          * Try to fill bufferOut from the message queue
          */
         private void processOut() {
-            while (!queue.isEmpty()) {
-                var message = queue.peek();
-                if (bufferOut.remaining() >= message.length(UTF8) + Integer.BYTES * 2) {
-                    message = queue.pop();
-                    bufferOut.putInt(message.login().getBytes(UTF_8).length).put(UTF8.encode(message.login())).putInt(message.msg().getBytes(UTF_8).length).put(UTF8.encode(message.msg()));
+            while (!requestQueue.isEmpty()) {
+                var request = requestQueue.peek();
+                if (bufferOut.remaining() >= request.length()) {
+                    request = requestQueue.pop();
+                    bufferOut.putInt(request.code().getOpCode()).put(request.buffer());
                 }
             }
         }
@@ -280,7 +329,7 @@ public class ServerChatFusion {
                 ops |= SelectionKey.OP_READ;
             }
 
-            if (!closed && (bufferOut.position() != 0 || !queue.isEmpty())) {
+            if (!closed && (bufferOut.position() != 0 || !requestQueue.isEmpty())) {
                 ops |= SelectionKey.OP_WRITE;
             }
 
@@ -297,8 +346,8 @@ public class ServerChatFusion {
                 silentlyClose();
             }
 
-            activeSinceLastTimeoutCheck = false;
-
+            // disabled for now
+            activeSinceLastTimeoutCheck = true;
         }
 
         private void silentlyClose() {
